Gestió dels accessos al backoffice #272 Modificat el tipus de PluginID

+++ carpeta-front/src/main/java/es/caib/carpeta/front/controller/PluginFrontController.java
@@ -244,7 +244,7 @@ public class PluginFrontController extends CommonFrontController {
             //ACCESS
 
             EntitatJPA entitatJPA = entitatEjb.findByCodi(sesionHttp.getEntitat());
-            accesLogicaEjb.crearAcces(usuarioClave,TIPUS_ACCES_PLUGIN, entitatJPA.getCodiDir3(),Integer.parseInt(pluginID),new Timestamp(new Date().getTime()),LocaleContextHolder.getLocale().getLanguage(), InetAddress.getLocalHost().getHostAddress());
+            accesLogicaEjb.crearAcces(usuarioClave,TIPUS_ACCES_PLUGIN, entitatJPA.getCodiDir3(),Long.parseLong(pluginID),new Timestamp(new Date().getTime()),LocaleContextHolder.getLocale().getLanguage(), InetAddress.getLocalHost().getHostAddress());
